move calculations classes to more organized location
