+++ src/app/calculations/analysis-calculations/annualAccountAnalysisSummaryClass.ts
@@ -3,8 +3,8 @@ import { CalanderizedMeter } from "src/app/models/calanderization";
 import { IdbAccount, IdbAccountAnalysisItem, IdbAnalysisItem, IdbFacility, IdbPredictorEntry } from "src/app/models/idb";
 import { AnnualAnalysisSummaryDataClass } from "./annualAnalysisSummaryDataClass";
 import { AnnualAnalysisSummary } from 'src/app/models/analysis';
-import { HelperService } from "./helperService";
 import { MonthlyAccountAnalysisClass } from "./monthlyAccountAnalysisClass";
+import { checkAnalysisValue } from "../shared-calculations/calculationsHelpers";
 
 export class AnnualAccountAnalysisSummaryClass {
 
@@ -12,7 +12,6 @@ export class AnnualAccountAnalysisSummaryClass {
     annualAnalysisSummaryDataClasses: Array<AnnualAnalysisSummaryDataClass>;
     baselineYear: number;
     reportYear: number;
-    helperService: HelperService;
     constructor(
         accountAnalysisItem: IdbAccountAnalysisItem,
         account: IdbAccount,
@@ -20,10 +19,9 @@ export class AnnualAccountAnalysisSummaryClass {
         accountFacilities: Array<IdbFacility>,
         accountPredictorEntries: Array<IdbPredictorEntry>,
         allAccountAnalysisItems: Array<IdbAnalysisItem>) {
-        this.helperService = new HelperService();
         this.setMonthlyAnalysisSummaryData(accountAnalysisItem, account, calanderizedMeters, accountFacilities, accountPredictorEntries, allAccountAnalysisItems);
         this.setBaselineYear(account);
-        this.setReportYear(accountAnalysisItem, account);
+        this.setReportYear(accountAnalysisItem);
         this.setAnnualAnalysisSummaryDataClasses(accountPredictorEntries);
     }
 
@@ -34,16 +32,10 @@ export class AnnualAccountAnalysisSummaryClass {
 
     setBaselineYear(account: IdbAccount) {
         this.baselineYear = account.sustainabilityQuestions.energyReductionBaselineYear;
-        // if (account.fiscalYear == 'nonCalendarYear' && account.fiscalYearCalendarEnd) {
-        //     this.baselineYear = this.baselineYear - 1;
-        // }
     }
 
-    setReportYear(analysisItem: IdbAccountAnalysisItem, account: IdbAccount) {
+    setReportYear(analysisItem: IdbAccountAnalysisItem) {
         this.reportYear = analysisItem.reportYear;
-        // if (account.fiscalYear == 'nonCalendarYear' && account.fiscalYearCalendarEnd) {
-        //     this.reportYear = this.reportYear - 1;
-        // }
     }
 
 
@@ -51,7 +43,6 @@ export class AnnualAccountAnalysisSummaryClass {
         this.annualAnalysisSummaryDataClasses = new Array();
         let analysisYear: number = this.baselineYear;
         while (analysisYear <= this.reportYear) {
-            // let annualAnalysisSummaryDataClassCopy: Array<AnnualAnalysisSummaryDataClass> = JSON.parse(JSON.stringify(this.annualAnalysisSummaryDataClasses))
             let yearAnalysisSummaryDataClass: AnnualAnalysisSummaryDataClass = new AnnualAnalysisSummaryDataClass(this.monthlyAnalysisSummaryData, analysisYear, accountPredictorEntries, undefined, this.annualAnalysisSummaryDataClasses);
             this.annualAnalysisSummaryDataClasses.push(yearAnalysisSummaryDataClass);
             analysisYear++;
@@ -69,12 +60,12 @@ export class AnnualAccountAnalysisSummaryClass {
                 baselineAdjustmentForNormalization: summaryDataClass.baselineAdjustmentForNormalization,
                 baselineAdjustmentForOther: summaryDataClass.baselineAdjustmentForOther,
                 baselineAdjustment: summaryDataClass.baselineAdjustment,
-                SEnPI: this.helperService.checkValue(summaryDataClass.SEnPI),
-                savings: this.helperService.checkValue(summaryDataClass.savings),
-                totalSavingsPercentImprovement: this.helperService.checkValue(summaryDataClass.totalSavingsPercentImprovement) * 100,
-                annualSavingsPercentImprovement: this.helperService.checkValue(summaryDataClass.annualSavingsPercentImprovement) * 100,
-                cummulativeSavings: this.helperService.checkValue(summaryDataClass.cummulativeSavings),
-                newSavings: this.helperService.checkValue(summaryDataClass.newSavings),
+                SEnPI: checkAnalysisValue(summaryDataClass.SEnPI),
+                savings: checkAnalysisValue(summaryDataClass.savings),
+                totalSavingsPercentImprovement: checkAnalysisValue(summaryDataClass.totalSavingsPercentImprovement) * 100,
+                annualSavingsPercentImprovement: checkAnalysisValue(summaryDataClass.annualSavingsPercentImprovement) * 100,
+                cummulativeSavings: checkAnalysisValue(summaryDataClass.cummulativeSavings),
+                newSavings: checkAnalysisValue(summaryDataClass.newSavings),
                 predictorUsage: summaryDataClass.predictorUsage
             }
         })
