+++ src/app/calculations/analysis-calculations/annualFacilityAnalysisSummaryClass.ts
@@ -3,8 +3,8 @@ import { CalanderizedMeter } from "src/app/models/calanderization";
 import { IdbAccount, IdbAnalysisItem, IdbFacility, IdbPredictorEntry } from "src/app/models/idb";
 import { AnnualAnalysisSummaryDataClass } from "./annualAnalysisSummaryDataClass";
 import { AnnualAnalysisSummary } from 'src/app/models/analysis';
-import { HelperService } from "./helperService";
 import { MonthlyFacilityAnalysisClass } from "./monthlyFacilityAnalysisClass";
+import { checkAnalysisValue } from "../shared-calculations/calculationsHelpers";
 
 export class AnnualFacilityAnalysisSummaryClass {
 
@@ -12,12 +12,10 @@ export class AnnualFacilityAnalysisSummaryClass {
     annualAnalysisSummaryDataClasses: Array<AnnualAnalysisSummaryDataClass>;
     baselineYear: number;
     reportYear: number;
-    helperService: HelperService;
     constructor(analysisItem: IdbAnalysisItem, facility: IdbFacility, calanderizedMeters: Array<CalanderizedMeter>, accountPredictorEntries: Array<IdbPredictorEntry>) {
-        this.helperService = new HelperService();
         this.setMonthlyAnalysisSummaryData(analysisItem, facility, calanderizedMeters, accountPredictorEntries);
         this.setBaselineYear(facility);
-        this.setReportYear(analysisItem, facility);
+        this.setReportYear(analysisItem);
         this.setAnnualAnalysisSummaryDataClasses(accountPredictorEntries, facility);
     }
 
@@ -28,16 +26,10 @@ export class AnnualFacilityAnalysisSummaryClass {
 
     setBaselineYear(facilityOrAccount: IdbFacility | IdbAccount) {
         this.baselineYear = facilityOrAccount.sustainabilityQuestions.energyReductionBaselineYear;
-        // if (facilityOrAccount.fiscalYear == 'nonCalendarYear' && facilityOrAccount.fiscalYearCalendarEnd) {
-        //     this.baselineYear = this.baselineYear - 1;
-        // }
     }
 
-    setReportYear(analysisItem: IdbAnalysisItem, facilityOrAccount: IdbFacility | IdbAccount) {
+    setReportYear(analysisItem: IdbAnalysisItem) {
         this.reportYear = analysisItem.reportYear;
-        // if (facilityOrAccount.fiscalYear == 'nonCalendarYear' && facilityOrAccount.fiscalYearCalendarEnd) {
-        //     this.reportYear = this.reportYear - 1;
-        // }
     }
 
 
@@ -45,7 +37,6 @@ export class AnnualFacilityAnalysisSummaryClass {
         this.annualAnalysisSummaryDataClasses = new Array();
         let analysisYear: number = this.baselineYear;
         while (analysisYear <= this.reportYear) {
-            // let annualAnalysisSummaryDataClassCopy: Array<AnnualAnalysisSummaryDataClass> = JSON.parse(JSON.stringify(this.annualAnalysisSummaryDataClasses))
             let yearAnalysisSummaryDataClass: AnnualAnalysisSummaryDataClass = new AnnualAnalysisSummaryDataClass(this.monthlyAnalysisSummaryData, analysisYear, accountPredictorEntries, facility, this.annualAnalysisSummaryDataClasses);
             this.annualAnalysisSummaryDataClasses.push(yearAnalysisSummaryDataClass);
             analysisYear++;
@@ -63,12 +54,12 @@ export class AnnualFacilityAnalysisSummaryClass {
                 baselineAdjustmentForNormalization: summaryDataClass.baselineAdjustmentForNormalization,
                 baselineAdjustmentForOther: summaryDataClass.baselineAdjustmentForOther,
                 baselineAdjustment: summaryDataClass.baselineAdjustment,
-                SEnPI: this.helperService.checkValue(summaryDataClass.SEnPI),
-                savings: this.helperService.checkValue(summaryDataClass.savings),
-                totalSavingsPercentImprovement: this.helperService.checkValue(summaryDataClass.totalSavingsPercentImprovement) * 100,
-                annualSavingsPercentImprovement: this.helperService.checkValue(summaryDataClass.annualSavingsPercentImprovement) * 100,
-                cummulativeSavings: this.helperService.checkValue(summaryDataClass.cummulativeSavings),
-                newSavings: this.helperService.checkValue(summaryDataClass.newSavings),
+                SEnPI: checkAnalysisValue(summaryDataClass.SEnPI),
+                savings: checkAnalysisValue(summaryDataClass.savings),
+                totalSavingsPercentImprovement: checkAnalysisValue(summaryDataClass.totalSavingsPercentImprovement) * 100,
+                annualSavingsPercentImprovement: checkAnalysisValue(summaryDataClass.annualSavingsPercentImprovement) * 100,
+                cummulativeSavings: checkAnalysisValue(summaryDataClass.cummulativeSavings),
+                newSavings: checkAnalysisValue(summaryDataClass.newSavings),
                 predictorUsage: summaryDataClass.predictorUsage
             }
         })
