+++ src/app/calculations/analysis-calculations/monthlyAnalysisCalculatedValuesClass.ts
