+++ src/app/calculations/analysis-calculations/monthlyFacilityAnalysisClass.ts
@@ -1,7 +1,8 @@
 import { MonthlyAnalysisSummaryData } from "src/app/models/analysis";
 import { CalanderizedMeter } from "src/app/models/calanderization";
 import { IdbAnalysisItem, IdbFacility, IdbPredictorEntry } from "src/app/models/idb";
-import { HelperService } from "./helperService";
+import { getFiscalYear } from "../shared-calculations/calanderizationFunctions";
+import { checkAnalysisValue, getMonthlyStartAndEndDate } from "../shared-calculations/calculationsHelpers";
 import { MonthlyAnalysisSummaryClass } from "./monthlyAnalysisSummaryClass";
 import { MonthlyAnalysisSummaryDataClass } from "./monthlyAnalysisSummaryDataClass";
 import { MonthlyFacilityAnalysisDataClass } from "./monthlyFacilityAnalysisDataClass";
@@ -12,11 +13,9 @@ export class MonthlyFacilityAnalysisClass {
     facilityMonthSummaries: Array<MonthlyFacilityAnalysisDataClass>;
     startDate: Date;
     endDate: Date;
-    helperService: HelperService;
     facilityPredictorEntries: Array<IdbPredictorEntry>;
     baselineYear: number;
     constructor(analysisItem: IdbAnalysisItem, facility: IdbFacility, calanderizedMeters: Array<CalanderizedMeter>, accountPredictorEntries: Array<IdbPredictorEntry>) {
-        this.helperService = new HelperService();
         this.setStartAndEndDate(facility, analysisItem);
         this.setGroupSummaries(analysisItem, facility, calanderizedMeters, accountPredictorEntries);
         this.setFacilityPredictorEntries(accountPredictorEntries, facility);
@@ -25,7 +24,7 @@ export class MonthlyFacilityAnalysisClass {
     }
 
     setStartAndEndDate(facility: IdbFacility, analysisItem: IdbAnalysisItem) {
-        let monthlyStartAndEndDate: { baselineDate: Date, endDate: Date } = this.helperService.getMonthlyStartAndEndDate(facility, analysisItem);
+        let monthlyStartAndEndDate: { baselineDate: Date, endDate: Date } = getMonthlyStartAndEndDate(facility, analysisItem);
         this.startDate = monthlyStartAndEndDate.baselineDate;
         this.endDate = monthlyStartAndEndDate.endDate;
     }
@@ -47,7 +46,7 @@ export class MonthlyFacilityAnalysisClass {
 
 
     setBaselineYear(facility: IdbFacility) {
-        this.baselineYear = this.helperService.getFiscalYear(this.startDate, facility);
+        this.baselineYear = getFiscalYear(this.startDate, facility);
     }
 
     setFacilityMonthSummaries(facility: IdbFacility) {
@@ -85,13 +84,13 @@ export class MonthlyFacilityAnalysisClass {
                 predictorUsage: summaryDataItem.predictorUsage,
                 fiscalYear: summaryDataItem.fiscalYear,
                 group: undefined,
-                SEnPI: this.helperService.checkValue(summaryDataItem.monthlyAnalysisCalculatedValues.SEnPI),
-                savings: this.helperService.checkValue(summaryDataItem.monthlyAnalysisCalculatedValues.savings),
-                percentSavingsComparedToBaseline: this.helperService.checkValue(summaryDataItem.monthlyAnalysisCalculatedValues.percentSavingsComparedToBaseline) * 100,
-                yearToDateSavings: this.helperService.checkValue(summaryDataItem.monthlyAnalysisCalculatedValues.yearToDateSavings),
-                yearToDatePercentSavings: this.helperService.checkValue(summaryDataItem.monthlyAnalysisCalculatedValues.yearToDatePercentSavings) * 100,
-                rollingSavings: this.helperService.checkValue(summaryDataItem.monthlyAnalysisCalculatedValues.rollingSavings),
-                rolling12MonthImprovement: this.helperService.checkValue(summaryDataItem.monthlyAnalysisCalculatedValues.rolling12MonthImprovement) * 100,
+                SEnPI: checkAnalysisValue(summaryDataItem.monthlyAnalysisCalculatedValues.SEnPI),
+                savings: checkAnalysisValue(summaryDataItem.monthlyAnalysisCalculatedValues.savings),
+                percentSavingsComparedToBaseline: checkAnalysisValue(summaryDataItem.monthlyAnalysisCalculatedValues.percentSavingsComparedToBaseline) * 100,
+                yearToDateSavings: checkAnalysisValue(summaryDataItem.monthlyAnalysisCalculatedValues.yearToDateSavings),
+                yearToDatePercentSavings: checkAnalysisValue(summaryDataItem.monthlyAnalysisCalculatedValues.yearToDatePercentSavings) * 100,
+                rollingSavings: checkAnalysisValue(summaryDataItem.monthlyAnalysisCalculatedValues.rollingSavings),
+                rolling12MonthImprovement: checkAnalysisValue(summaryDataItem.monthlyAnalysisCalculatedValues.rolling12MonthImprovement) * 100,
             }
         })
     }
