+++ src/app/calculations/analysis-calculations/monthlyFacilityAnalysisDataClass.ts
@@ -1,37 +1,43 @@
-import { IdbAccount, IdbAccountAnalysisItem } from "src/app/models/idb";
-import { HelperService } from "./helperService";
+import { IdbFacility, IdbPredictorEntry, PredictorData } from "src/app/models/idb";
+import * as _ from 'lodash';
 import { MonthlyAnalysisCalculatedValues } from "./monthlyAnalysisCalculatedValuesClass";
 import { MonthlyAnalysisSummaryDataClass } from "./monthlyAnalysisSummaryDataClass";
-import * as _ from 'lodash';
-export class MonthlyAccountAnalysisDataClass {
+import { getFiscalYear } from "../shared-calculations/calanderizationFunctions";
 
+export class MonthlyFacilityAnalysisDataClass {
 
     date: Date;
     energyUse: number;
     modeledEnergy: number;
     baselineAdjustmentForOther: number;
+    predictorUsage: Array<{
+        usage: number,
+        predictorId: string
+    }>;
     fiscalYear: number;
     monthlyAnalysisCalculatedValues: MonthlyAnalysisCalculatedValues;
 
     currentMonthData: Array<MonthlyAnalysisSummaryDataClass>;
+    monthPredictorData: Array<IdbPredictorEntry>;
     baselineActualEnergyUse: number;
     monthIndex: number;
-    helperService: HelperService;
+    facilityGUID: string;
     constructor(
         allFacilityAnalysisData: Array<MonthlyAnalysisSummaryDataClass>,
         monthDate: Date,
-        previousMonthsSummaryData: Array<MonthlyAccountAnalysisDataClass>,
+        facilityPredictorEntries: Array<IdbPredictorEntry>,
+        previousMonthsSummaryData: Array<MonthlyFacilityAnalysisDataClass>,
         baselineYear: number,
-        account: IdbAccount,
-        analysisItem: IdbAccountAnalysisItem,
-        annualUsageValues: Array<{year: number, usage: number}>) {
-        this.helperService = new HelperService();
+        facility: IdbFacility) {
+        this.facilityGUID = facility.guid;
         this.date = monthDate;
-        this.setFiscalYear(account);
+        this.setFiscalYear(facility);
         this.setCurrentMonthData(allFacilityAnalysisData);
+        this.setMonthPredictorData(facilityPredictorEntries);
+        this.setPredictorUsage(facilityPredictorEntries);
         this.setEnergyUse();
         this.setModeledEnergy();
-        this.setBaselineAdjustmentForOther(analysisItem, baselineYear, annualUsageValues);
+        this.setBaselineAdjustmentForOther();
         this.setMonthIndex(previousMonthsSummaryData);
         this.setBaselineActualEnergyUse(baselineYear, previousMonthsSummaryData);
         this.setMonthlyAnalysisCalculatedValues(baselineYear, previousMonthsSummaryData);
@@ -44,8 +50,33 @@ export class MonthlyAccountAnalysisDataClass {
         });
     }
 
-    setFiscalYear(facility: IdbAccount) {
-        this.fiscalYear = this.helperService.getFiscalYear(new Date(this.date), facility);
+    setMonthPredictorData(facilityPredictorEntries: Array<IdbPredictorEntry>) {
+        this.monthPredictorData = facilityPredictorEntries.filter(predictorData => {
+            let predictorDate: Date = new Date(predictorData.date);
+            return predictorDate.getUTCFullYear() == this.date.getUTCFullYear() && predictorDate.getUTCMonth() == this.date.getUTCMonth();
+        });
+    }
+
+    setPredictorUsage(facilityPredictorEntries: Array<IdbPredictorEntry>) {
+        this.predictorUsage = new Array();
+        if (facilityPredictorEntries.length != 0) {
+            let predictorVariables: Array<PredictorData> = facilityPredictorEntries[0].predictors;
+            predictorVariables.forEach(variable => {
+                let usageVal: number = 0;
+                this.monthPredictorData.forEach(data => {
+                    let predictorData: PredictorData = data.predictors.find(predictor => { return predictor.id == variable.id });
+                    usageVal = usageVal + predictorData.amount;
+                });
+                this.predictorUsage.push({
+                    usage: usageVal,
+                    predictorId: variable.id
+                });
+            });
+        }
+    }
+
+    setFiscalYear(facility: IdbFacility) {
+        this.fiscalYear = getFiscalYear(new Date(this.date), facility);
     }
 
     setEnergyUse() {
@@ -56,24 +87,16 @@ export class MonthlyAccountAnalysisDataClass {
         this.modeledEnergy = _.sumBy(this.currentMonthData, 'modeledEnergy');
     }
 
-    setBaselineAdjustmentForOther(accountAnalysisItem: IdbAccountAnalysisItem, baselineYear: number, annualUsageValues: Array<{year: number, usage: number}>) {
+    setBaselineAdjustmentForOther() {
         this.baselineAdjustmentForOther = _.sumBy(this.currentMonthData, 'baselineAdjustmentForOther');
-        if (accountAnalysisItem.hasBaselineAdjustement && this.fiscalYear != baselineYear) {
-            let annualEnergyUse: number = annualUsageValues.find(usageVal => {return usageVal.year == this.fiscalYear})?.usage;
-            let yearAdjustment: { year: number, amount: number } = accountAnalysisItem.baselineAdjustments.find(bAdjustement => { return bAdjustement.year == this.date.getUTCFullYear(); })
-            if (yearAdjustment.amount) {
-                let accountBaselineAdjustementForOther: number = (this.energyUse / annualEnergyUse) * yearAdjustment.amount;
-                this.baselineAdjustmentForOther = this.baselineAdjustmentForOther + accountBaselineAdjustementForOther;
-            }
-        }
     }
 
-    setMonthIndex(previousMonthsSummaryData: Array<MonthlyAccountAnalysisDataClass>) {
+    setMonthIndex(previousMonthsSummaryData: Array<MonthlyFacilityAnalysisDataClass>) {
         let summaryDataIndex: number = previousMonthsSummaryData.length;
         if (summaryDataIndex == 0) {
             this.monthIndex = 0;
         } else {
-            let previousMonthSummaryData: MonthlyAccountAnalysisDataClass = previousMonthsSummaryData[summaryDataIndex - 1];
+            let previousMonthSummaryData: MonthlyFacilityAnalysisDataClass = previousMonthsSummaryData[summaryDataIndex - 1];
             if (previousMonthSummaryData.fiscalYear == this.fiscalYear) {
                 this.monthIndex = previousMonthSummaryData.monthIndex + 1;
             } else {
@@ -82,14 +105,14 @@ export class MonthlyAccountAnalysisDataClass {
         }
     }
 
-    setBaselineActualEnergyUse(baselineYear: number, previousMonthsSummaryData: Array<MonthlyAccountAnalysisDataClass>) {
+    setBaselineActualEnergyUse(baselineYear: number, previousMonthsSummaryData: Array<MonthlyFacilityAnalysisDataClass>) {
         if (this.fiscalYear == baselineYear) {
             this.baselineActualEnergyUse = this.energyUse;
         } else {
             this.baselineActualEnergyUse = previousMonthsSummaryData[this.monthIndex].energyUse;
         }
     }
-    setMonthlyAnalysisCalculatedValues(baselineYear: number, previousMonthsSummaryData: Array<MonthlyAccountAnalysisDataClass>) {
+    setMonthlyAnalysisCalculatedValues(baselineYear: number, previousMonthsSummaryData: Array<MonthlyFacilityAnalysisDataClass>) {
         let previousMonthsAnalysisCalculatedValues: Array<MonthlyAnalysisCalculatedValues> = previousMonthsSummaryData.map(data => { return data.monthlyAnalysisCalculatedValues });
         this.monthlyAnalysisCalculatedValues = new MonthlyAnalysisCalculatedValues(
             this.energyUse,
