+++ src/app/calculations/better-plants-calculations/betterPlantsEnergySummaryClass.ts
