+++ src/app/calculations/better-plants-calculations/betterPlantsReportClass.ts
@@ -2,8 +2,8 @@ import { AnnualAnalysisSummary } from "src/app/models/analysis";
 import { CalanderizedMeter } from "src/app/models/calanderization";
 import { IdbAccount, IdbAccountAnalysisItem, IdbAnalysisItem, IdbFacility, IdbPredictorEntry } from "src/app/models/idb";
 import { BetterPlantsSummary, ReportOptions } from "src/app/models/overview-report";
-import { AnnualAccountAnalysisSummaryClass } from "./annualAccountAnalysisSummaryClass";
-import { AnnualFacilityAnalysisSummaryClass } from "./annualFacilityAnalysisSummaryClass";
+import { AnnualAccountAnalysisSummaryClass } from "src/app/calculations/analysis-calculations/annualAccountAnalysisSummaryClass";
+import { AnnualFacilityAnalysisSummaryClass } from "src/app/calculations/analysis-calculations/annualFacilityAnalysisSummaryClass";
 import { BetterPlantsEnergySummaryClass } from "./betterPlantsEnergySummaryClass";
 
 export class BetterPlantsReportClass {
