+++ src/app/web-workers/account-overview.worker.ts
@@ -1,6 +1,6 @@
 /// <reference lib="webworker" />
 
-import { AccountSummaryClass } from "./classes/dashboard/accountSummaryClass";
+import { AccountSummaryClass } from "../calculations/dashboard-calculations/accountSummaryClass";
 
 addEventListener('message', ({ data }) => {
     let accountFacilitiesSummaryClass: AccountSummaryClass = new AccountSummaryClass(data.calanderizedMeters, data.facilities, data.sources, data.account);
