+++ src/app/web-workers/annual-account-analysis.worker.ts
@@ -1,7 +1,7 @@
 /// <reference lib="webworker" />
 
 import { AnnualAnalysisSummary, MonthlyAnalysisSummaryData } from "../models/analysis";
-import { AnnualAccountAnalysisSummaryClass } from "./classes/annualAccountAnalysisSummaryClass";
+import { AnnualAccountAnalysisSummaryClass } from "../calculations/analysis-calculations/annualAccountAnalysisSummaryClass";
 
 addEventListener('message', ({ data }) => {
     let annualAnalysisSummaryClass: AnnualAccountAnalysisSummaryClass = new AnnualAccountAnalysisSummaryClass(data.accountAnalysisItem, data.account, data.calanderizedMeters, data.accountFacilities, data.accountPredictorEntries, data.allAccountAnalysisItems);
