+++ src/app/web-workers/annual-group-analysis.worker.ts
@@ -1,7 +1,7 @@
 /// <reference lib="webworker" />
 
 import { AnnualAnalysisSummary } from "src/app/models/analysis";
-import { AnnualGroupAnalysisSummaryClass } from "./classes/annualGroupAnalysisSummaryClass";
+import { AnnualGroupAnalysisSummaryClass } from "src/app/calculations/analysis-calculations/annualGroupAnalysisSummaryClass";
 
 addEventListener('message', ({ data }) => {
     let annualAnalysisSummaryClass: AnnualGroupAnalysisSummaryClass = new AnnualGroupAnalysisSummaryClass(data.selectedGroup, data.analysisItem, data.facility, data.calanderizedMeters, data.accountPredictorEntries);
