+++ src/app/web-workers/monthly-group-analysis.worker.ts
@@ -1,7 +1,7 @@
 /// <reference lib="webworker" />
 
 import { MonthlyAnalysisSummary, MonthlyAnalysisSummaryData } from "../models/analysis";
-import { MonthlyAnalysisSummaryClass } from "./classes/monthlyAnalysisSummaryClass";
+import { MonthlyAnalysisSummaryClass } from "src/app/calculations/analysis-calculations/monthlyAnalysisSummaryClass";
 
 addEventListener('message', ({ data }) => {
   let monthlyAnalysisSummaryClass: MonthlyAnalysisSummaryClass = new MonthlyAnalysisSummaryClass(data.selectedGroup, data.analysisItem, data.facility, data.calanderizedMeters, data.accountPredictorEntries);
